[syle] : common 클래스 패키지 이동으로 인한 단순변경
