--- src/main/java/com/core/gtd/common/response/ResponseStatus.java
@@ -1,32 +1,0 @@
-package com.core.gtd.common.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-import static org.springframework.http.HttpStatus.*;
-
-@Getter
-@AllArgsConstructor
-public enum ResponseStatus {
-
-    /**
-     * 200(OK)
-     */
-    SUCCESS(true, OK.value(), "요청에 성공했습니다."),
-
-    /**
-     * 400(Bad_request)
-     */
-
-    /**
-     * 500(Internal_Server_Error)
-     */
-    INVALID_ERROR(false, INTERNAL_SERVER_ERROR.value(), "예상치 못한 에러가 발생했습니다.");
-
-    private final boolean isSuccess;
-    private final int statusCode;
-    private final String message;
-
-}
-
