+++ src/main/java/com/core/gtd/src/basket/controller/TaskDetailController.java
@@ -1,6 +1,6 @@
 package com.core.gtd.src.basket.controller;
 
-import com.core.gtd.common.response.BaseResponse;
+import com.core.gtd.src.common.response.BaseResponse;
 import com.core.gtd.src.basket.model.dto.TaskDetailDto;
 import com.core.gtd.src.basket.model.request.DetailCompleteRequest;
 import com.core.gtd.src.basket.model.request.TaskDetailRequest;
@@ -8,10 +8,9 @@ import com.core.gtd.src.basket.model.response.TaskDetailResponse;
 import com.core.gtd.src.basket.service.TaskDetailService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
-import static com.core.gtd.common.response.BaseResponse.response;
+import static com.core.gtd.src.common.response.BaseResponse.response;
 
 @Slf4j
 @RestController
