+++ src/main/java/com/core/gtd/src/basket/model/dto/TaskDto.java
@@ -1,11 +1,10 @@
 package com.core.gtd.src.basket.model.dto;
 
-import com.core.gtd.common.constatnt.Priority;
-import com.core.gtd.common.constatnt.State;
-import com.core.gtd.common.constatnt.TaskState;
+import com.core.gtd.src.common.constatnt.Priority;
+import com.core.gtd.src.common.constatnt.State;
+import com.core.gtd.src.common.constatnt.TaskState;
 import com.core.gtd.src.basket.entity.Task;
-import com.core.gtd.src.basket.entity.TaskDetail;
-import com.core.gtd.src.basket.model.response.TaskDetailResponse;
+import com.core.gtd.src.users.model.dto.UserDto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -13,9 +12,7 @@ import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Getter
@@ -35,6 +32,7 @@ public class TaskDto {
     private LocalDateTime startAt;
     private LocalDateTime deadlineAt;
     private List<TaskDetailDto> taskDetail = new ArrayList<>();
+    private UserDto userDto;
 
     public static TaskDto fromEntity(Task task) {
         return TaskDto.builder()
@@ -48,11 +46,11 @@ public class TaskDto {
                 .updatedAt(task.getUpdatedAt())
                 .startAt(task.getStartAt())
                 .deadlineAt(task.getDeadlineAt())
-                .taskDetail(task.getTaskDetail() != null ?
-                        task.getTaskDetail().stream()
+                .taskDetail(task.getTaskDetail().stream()
                                 .map(TaskDetailDto::fromEntity)
-                                .collect(Collectors.toList()) :
-                        Collections.emptyList())
+                                .collect(Collectors.toList()))
+                .userDto(UserDto.fromEntity(task.getUser()))
                 .build();
+
     }
 }
