+++ src/main/java/com/core/gtd/src/basket/model/request/DetailCompleteRequest.java
@@ -1,6 +1,6 @@
 package com.core.gtd.src.basket.model.request;
 
-import com.core.gtd.common.constatnt.IsComplete;
+import com.core.gtd.src.common.constatnt.IsComplete;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
