+++ src/main/java/com/core/gtd/src/basket/model/request/TaskDetailRequest.java
@@ -1,7 +1,7 @@
 package com.core.gtd.src.basket.model.request;
 
-import com.core.gtd.common.constatnt.IsComplete;
-import com.core.gtd.common.constatnt.State;
+import com.core.gtd.src.common.constatnt.IsComplete;
+import com.core.gtd.src.common.constatnt.State;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
