+++ src/main/java/com/core/gtd/src/common/exception/GlobalExceptionHandler.java
@@ -1,7 +1,7 @@
-package com.core.gtd.common.exception;
+package com.core.gtd.src.common.exception;
 
-import com.core.gtd.common.response.BaseResponse;
-import com.core.gtd.common.response.ResponseStatus;
+import com.core.gtd.src.common.response.BaseResponse;
+import com.core.gtd.src.common.response.ResponseStatus;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -23,6 +23,7 @@ public class GlobalExceptionHandler {
     @ExceptionHandler(Exception.class)
     public ResponseEntity<BaseResponse> exceptionHandler(Exception e) {
         log.error("[InternalServerError Occurs] error: {}", e.getMessage());
+        e.printStackTrace();
 
         return ResponseEntity.internalServerError()
                 .body(BaseResponse.response(ResponseStatus.INVALID_ERROR));
